% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_homeranges.R
\name{AIC.RhrParamEst}
\alias{AIC.RhrParamEst}
\title{Obtain the AIC of a home-range}
\usage{
\method{AIC}{RhrParamEst}(x, AICc = FALSE)
}
\arguments{
\item{x}{The home-range estimate.}

\item{AICc}{Logical value, indicating if the AICc should be returned instead of the AIC.}
}
\description{
Function to obtain the AIC of a home range. This is currently only possible for parametric home-ranges.
}

