\name{alert}
\alias{alert}
\title{#   <a class="accordion-toggle" data-toggle="collapse" data-parent=", id, '" href="#', gsub(" ", "", x[[1]]), '">', x[[1]],
alert}
\usage{
  alert(x, cat = TRUE)
}
\arguments{
  \item{x}{a string}

  \item{cat}{whether it should be printed to sto or not}
}
\description{
  wraps string as bootstrap alert
}

