\name{alertSuccess}
\alias{alertSuccess}
\title{alertSuccess}
\usage{
  alertSuccess(x, cat = TRUE)
}
\arguments{
  \item{x}{a string}

  \item{cat}{whether it should be printed to sto or not}
}
\description{
  wraps string as bootstrap alert-success
}
\examples{
alertSuccess("Everything went good")
}
\author{
  Johannes Signer
}

