\name{alertWarning}
\alias{alertWarning}
\title{alertWarning}
\usage{
  alertWarning(x, cat = TRUE)

  alertWarning(x, cat = TRUE)
}
\arguments{
  \item{x}{a string}

  \item{cat}{whether it should be printed to sto or not}

  \item{x}{a string}

  \item{cat}{whether it should be printed to sto or not}
}
\description{
  wraps string as bootstrap alert-warning

  wraps string as bootstrap alert-success
}

