\name{areas.RhrHREstimator}
\alias{areas.RhrHREstimator}
\title{Get areas from rhrHREstimator}
\usage{
  \method{areas}{RhrHREstimator} (x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\value{
  data.frame
}
\description{
  Get areas from rhrHREstimator
}

