\name{cud}
\alias{cud}
\title{Retrives the CUD of an rhrHREstimator object}
\usage{
  cud(x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\value{
  the UD raster
}
\description{
  Retrives the CUD of an rhrHREstimator object
}

