\name{cud.RhrHREstimator}
\alias{cud.RhrHREstimator}
\title{Retrives the CUD of an rhrHREstimator object}
\usage{
  \method{ud}{RhrHREstimator} (x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\description{
  Retrives the CUD of an rhrHREstimator object
}

