% Generated by roxygen2 (4.0.0): do not edit by hand
\name{expandCall}
\alias{expandCall}
\title{expandCall}
\usage{
expandCall(call, n = 2)
}
\arguments{
  \item{call}{a call}

  \item{n}{integer, the number of generations to go back in
  the parent frame}
}
\value{
call
}
\description{
Generates a call with all arguments as literals
}
\details{
This is a slightly adapted form from:
http://stackoverflow.com/questions/3478923/displaying-the-actual-parameter-list-of-the-function-during-execution
}
\author{
Johannes Signer
}

