\name{fidelityBase}
\alias{fidelityBase}
\title{Base function to calculate msd and li}
\usage{
  fidelityBase(x, y, n)
}
\arguments{
  \item{x}{x-coordinates}

  \item{y}{y-coordinates}

  \item{n}{the number of coordinates}
}
\value{
  a list
}
\description{
  Base function to calculate msd and li
}
\author{
  Johannes Signer
}

