\name{hasCUD.RhrHREstimator}
\alias{hasCUD.RhrHREstimator}
\title{Checks if an rhrEstimator posses a cud}
\usage{
  \method{hasCUD}{RhrHREstimator} (x)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}
}
\value{
  Returns \code{TRUE} if the \code{rhrHREstimator} has a
  utilisation distribution, else it returns false.
}
\description{
  Checks if an rhrEstimator posses a cud
}

