\name{hasIsopleths.RhrHREstimator}
\alias{hasIsopleths.RhrHREstimator}
\title{Check if isopleths are available}
\usage{
  \method{hasIsopleths}{RhrHREstimator} (x)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}
}
\value{
  Returns \code{TRUE} if the \code{rhrHREstimator} has one
  or more isopleths, else it returns false.
}
\description{
  Check if isopleths are available
}

