% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrArgs.R
\name{hr_utility}
\alias{hr_utility}
\alias{rhrArgs}
\title{Functions to work with results of a home range estimate.}
\usage{
rhrArgs(x, ...)
}
\arguments{
\item{x}{An object of class RhrEst*.}

\item{...}{None implemented.}
}
\value{
A list.
}
\description{
This collection is a set of utility function to work with results of a home range estimate. \code{rhrArgs} returns the arguments that where passed to a call.
}
\details{
More to do here

Returns a list of all arguments that were passed to an \code{rhr*} function call
}
\examples{
data(datSH)
mcp <- rhrMCP(datSH[, 2:3])
str(rhrArgs(mcp))
}

