% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{is.regular}
\alias{is.regular}
\title{Checks if the track is regularly spaced in time.}
\usage{
is.regular(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrackS*}}

\item{...}{none implemented.}
}
\description{
Returns `TRUE` if time differences between consecutive relocations are equal.
}
\examples{
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Create an object of RhrTrackS (only space)
trackS <- rhrTrack(sp)

is.regular(trackS)
}

