\name{isopleths}
\alias{isopleths}
\title{Retrives the isopleths of an rhrHREstimator object}
\usage{
  isopleths(x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\value{
  SpatialPolygonsDataFrame
}
\description{
  Retrives the isopleths of an rhrHREstimator object
}

