\name{isopleths.RhrHREstimator}
\alias{isopleths.RhrHREstimator}
\title{get the isopleths from rhrHREstimator}
\usage{
  \method{isopleths}{RhrHREstimator} (x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\description{
  get the isopleths from rhrHREstimator
}

