% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrOverlaps.R
\name{overlaps}
\alias{overlaps}
\alias{rhrBA}
\alias{rhrOverlap}
\title{Different Methods to calculate home-range overlaps (in alpha)}
\usage{
rhrOverlap(x, ...)

rhrBA(x, ...)
}
\arguments{
\item{x}{RhrEst}

\item{...}{None implemented.}

\item{y}{RhrEst}
}
\value{
\code{data.frame} with the isopleth level and area in units of the coordinate system.
}
\description{
Different Methods to calculate home-range overlaps (in alpha)
}

