\name{parameters.RhrHREstimator}
\alias{parameters.RhrHREstimator}
\title{Retrieves parameters of an object of class RhrHREstimator}
\usage{
  \method{parameters}{RhrHREstimator} (x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\value{
  List
}
\description{
  Retrieves parameters of an object of class RhrHREstimator
}

