% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_homeranges.R
\name{parametric_homeranges}
\alias{parametric_homeranges}
\alias{rhrBiCirc}
\alias{rhrBiNorm}
\alias{rhrUniCirc}
\alias{rhrUniNorm}
\title{rhrBiCirc}
\usage{
rhrUniCirc(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL), levels = 95)

rhrUniNorm(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL), levels = 95)

rhrBiCirc(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL))

rhrBiNorm(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL), maxit = 20)
}
\arguments{
\item{xy}{Instance of \code{data.frame}, \code{SpatialPoints},
\code{SpatialPointsDataFrame} or \code{RhrMappedData}. If \code{xy} is
is a \code{data.frame}, then the first two columns are assumed to be the
\code{x} and \code{y} coordinates of the relocations.}

\item{trast}{Instance of \code{RasterLayer} that serves as a template for the resulting grid of the estimation.}

\item{maxit}{Integer, giving the maximum number of iterations. See also \link{\code{mixtools::mvnormalmixEM}}, which is called.}
}
\description{
Bimodal circular home range estimation
}
\details{
Computes home-range using two circular normal distributions
}

