\name{plot.RhrHRAsymptote}
\alias{plot}
\alias{plot.RhrHRAsymptote}
\title{plot for RhrHRAsymptote}
\usage{
  plot(x, ...)
}
\arguments{
  \item{x}{RhrHRAsymptote}

  \item{draw}{indicates whether the plot should be drawn or
  not. If this is \code{FALSE} a grob is returned.}

  \item{...}{none implemented}
}
\description{
  generic plot for RhrHREstimator
}

