% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.RhrHRCoreArea}
\alias{plot.RhrHRCoreArea}
\title{plot for RhrHRCoreArea}
\usage{
\method{plot}{RhrHRCoreArea}(x, ...)
}
\arguments{
  \item{x}{RhrHRCoreArea object}

  \item{...}{none implemented}
}
\description{
generic plot for RhrHREstimator
}

