% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{plot.RhrTracks}
\alias{plot.RhrTracks}
\title{Prints RhrTracks object.}
\usage{
\method{plot}{RhrTracks}(x, ...)
}
\arguments{
\item{x}{RhrTracks object.}

\item{...}{None implemented.}
}
\description{
Prints RhrTracks object.
}

