% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{plotTime}
\alias{plotTime}
\title{Plot time}
\usage{
plotTime(x, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}

\item{...}{none implemented.}
}
\description{
Plot a histogram of times when relocations where recorded
}

