\name{plotTTSI}
\alias{plotTTSI}
\title{plot ttsi}
\usage{
  plotTTSI(res, cvReached, cvReachedAt)
}
\arguments{
  \item{res}{from ttsi}

  \item{cvReached}{whether or not the critical value was
  passed}

  \item{cvReachedAt}{time interval at which the critical
  value was reached}
}
\description{
  plots results of rhrSchoener
}

