\name{print.RhrHRAsymptote}
\alias{print}
\alias{print.RhrHRAsymptote}
\title{print object of rhrHRAsymptote}
\usage{
  print(x, ...)
}
\arguments{
  \item{x}{RhrHRAsymptote}

  \item{how}{should it be printed, this can be either
  sceen, html or grob}
}
\description{
  print object of rhrHRAsymptote
}

