\name{print.RhrHREstimator}
\alias{print}
\alias{print.RhrHREstimator}
\title{print for RhrHREstimator}
\usage{
  print(x, ...)
}
\arguments{
  \item{x}{RhrHREstimator object}

  \item{how}{how the output should be formated, can be
  either screen, html or grob}

  \item{...}{ignored}
}
\description{
  generic print for RhrHREstimator
}

