\name{printSpRestriction}
\alias{printSpRestriction}
\title{printSpRestriction}
\usage{
  printSpRestriction(datrm, datsub, latmin, latmax, lonmin,
    lonmax)
}
\arguments{
  \item{datrm}{original dataset}

  \item{datsub}{subsetted dataset}

  \item{latmin}{min y}

  \item{latmax}{max y}

  \item{lonmin}{min x}

  \item{lonmax}{max x}
}
\description{
  helper function for the gui
}

