\name{printTemporalRestriction}
\alias{printTemporalRestriction}
\title{printTemporalRestriction}
\usage{
  printTemporalRestriction(datrm, datsub, tmin, tmax)
}
\arguments{
  \item{datrm}{original dataset}

  \item{datsub}{subsetted dataset}

  \item{tmin}{min time}

  \item{tmax}{max time}
}
\description{
  helper function for the gui
}

