% Generated by roxygen2 (4.0.0): do not edit by hand
\name{progressInit}
\alias{progressInit}
\title{Initialize progress}
\usage{
progressInit()
}
\description{
Call this function in your \code{shinyUI} definition if you
intend to use progress in \code{server.R}.
}
\seealso{
\code{\link{withProgress}}, \code{\link{Progress}}
}

