\name{rasterFromXYVect}
\alias{rasterFromXYVect}
\title{rasterFromXYVect}
\usage{
  rasterFromXYVect(xy, xrange = NA, yrange = NA, res = 100)
}
\arguments{
  \item{xy}{a data.frame or matrix. The first column are x
  coordinates and the second column are y coordinates}

  \item{xrange}{range of x}

  \item{yrange}{range of y}

  \item{res}{resolution}
}
\description{
  Creates empty raster
}

