% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhr.R
\docType{package}
\name{rhr}
\alias{rhr}
\alias{rhr-package}
\title{rhr: A package for analyzing animal movement data}
\description{
The rhr (reproducable home ranges) provides function to
handle and analyze animal biotelemetry data. The rhr package
provides many functions from the commandline and throught a
graphical user interface (GUI).
}
\section{Functions}{
  The package implements common methods to estimate home
  ranges. Currently available methods are: \itemize{ \item
  Something important \itme Something even more important }
}

