% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrAIC.R
\name{rhrAIC}
\alias{rhrAIC}
\title{Obtain the AIC of a parametric home-range}
\usage{
rhrAIC(x, AICc = FALSE, ...)
}
\arguments{
\item{x}{The home-range estimate.}

\item{AICc}{Logical value, indicating if the AICc should be returned instead of the AIC.}

\item{...}{None implemented.}
}
\description{
Generic function to obtain Akaike's Information Criterion (AIC) for parametric home-ranges. This is currently only possible for parametric home-ranges.
}
\examples{
# Simulated data
set.seed(123)
dat <- data.frame(x = c(rnorm(100), rnorm(100, 20)), 
y = c(rnorm(100), rnorm(100, 20)))
uniCir <- rhrUniCirc(dat)
AIC(uniCir)
AIC(uniCir, AICc = TRUE)
biCir <- rhrBiCirc(dat)
AIC(biCir)
AIC(biCir, AICc = TRUE)
uniNorm <- rhrUniNorm(dat)
AIC(uniNorm)
AIC(uniNorm, AICc = TRUE)
biNorm <- rhrBiNorm(dat)
AIC(biNorm)
AIC(biNorm, AICc = TRUE)
}
\seealso{
\code{\link[stats]{AIC}}
}

