\name{rhrAlert}
\alias{rhrAlert}
\title{rhrAlert
Wraps a character as html \code{div} with appropriate alert classe to use with twitter bootstrap.}
\usage{
  rhrAlert(x, cat = TRUE, class = "info")
}
\arguments{
  \item{x}{a character}

  \item{cat}{logical, if true results will be \code{cat} to
  the command line}

  \item{class}{character, giving finer specification of the
  class, possible values are: error, info, success or
  warning.}
}
\value{
  \code{character}
}
\description{
  rhrAlert Wraps a character as html \code{div} with
  appropriate alert classe to use with twitter bootstrap.
}

