\name{rhrArea.KDE}
\alias{rhrArea.KDE}
\title{Area for Kernel Density Estimation}
\usage{
  rhrArea.KDE(est, level)
}
\arguments{
  \item{xy}{An object of class
  \code{SpatialPointsDataFrame}}

  \item{level}{The level of the UD at which area is
  calculated}

  \item{...}{passed to KDE}
}
\value{
  The area
}
\description{
  Area for Kernel Density Estimation
}

