\name{rhrArea}
\alias{rhrArea}
\alias{rhrArea.RhrHREstimator}
\title{Get areas from object of class RhrHREstimator}
\usage{
  rhrArea(x, ...)

  \method{rhrArea}{RhrHREstimator} (x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{...}{further arguments, none implemented}
}
\value{
  \code{data.frame}
}
\description{
  Get areas from object of class RhrHREstimator
}

