\name{rhrArea.RhrEstimatorMCP}
\alias{rhrArea.RhrEstimatorMCP}
\title{Area for Minimum Convex Polygon}
\usage{
  \method{rhrArea}{RhrEstimatorMCP} (x, ...)
}
\arguments{
  \item{x}{a RhrEstimatorMCP object}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Area for Minimum Convex Polygon
}

