\name{rhrAreaMCP}
\alias{rhrAreaMCP}
\title{Area for Minimum Convex Polygon}
\usage{
  rhrAreaMCP(xy, level)
}
\arguments{
  \item{xy}{An object of class
  \code{SpatialPointsDataFrame}}

  \item{level}{The level at which MCP is calculated}
}
\value{
  The area
}
\description{
  Area for Minimum Convex Polygon
}

