% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrArgs.R
\name{rhrArgs}
\alias{rhrArgs}
\title{Arguments passed to a function call}
\usage{
rhrArgs(x, ...)
}
\arguments{
\item{x}{An object of class RhrEst*.}

\item{...}{none implemented.}
}
\value{
A list.
}
\description{
Returns a list of all arguments that were passed to an \code{rhr*} function call
}
\examples{
library(rhr)
data(datSH)

## Geometric estimator
mcp <- rhrMCP(datSH[, 2:3])
rhrArgs(mcp)
}

