% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrCPPFunctions.R
\name{rhrBaseLSCV}
\alias{rhrBaseLSCV}
\title{Least Square Cross Validation}
\usage{
rhrBaseLSCV(x, y, h)
}
\arguments{
\item{x}{Numeric vector, the x coordinates.}

\item{y}{Numeric vector, the y cooridnates.}

\item{h}{Numeric vector, candidate bandwidth.}
}
\value{
Numeric vector, the score value for each of the proposed candidate bandwidths.
}
\description{
Least square cross validation to determine bandwidth for  kernel density estimation.
}
\references{
\url{http://www.esajournals.org/doi/abs/10.2307/2265701}
}

