% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrCPPFunctions.R
\name{rhrBaseMSD}
\alias{rhrBaseMSD}
\title{Mean squared distance}
\usage{
rhrBaseMSD(x, y)
}
\arguments{
\item{x}{Numeric vector, the x cooridnates}

\item{y}{Numeric vector, the y coordinates}
}
\value{
Numeric matrix with the x and y coordinates of the new path.
}
\description{
Mean squared distance from the center of activity.
}
\references{
\url{http://www.esajournals.org/doi/abs/10.2307/1937590}
}

