% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrBBMM.R
\name{rhrBBMM}
\alias{rhrBBMM}
\title{Brownian Bridge Movement Model (BBMM)}
\usage{
rhrBBMM(xy, rangesigma1 = c(0, 10000), sigma2 = 100,
  trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange = 0.2), nrow =
  100, res = NULL))
}
\arguments{
\item{rangesigma1}{parameter 1}

\item{sigma2}{parameter 2}

\item{trast}{a \code{RasterLayer} used as an template for the output grid.}

\item{track}{Object inheriting from \code{RhrTrackST}.}
}
\value{
object of class \code{RhrBBMM}
}
\description{
A wrapper around \code{adehabitatHR::kernelbb}.
}
\seealso{
\code{\link{rhrTrack}} for fruther details and examples on how to create a \code{RhrTrack*} object.

\code{adehabitatHR::kernelbb}
}

