% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrBBX}
\alias{rhrBBX}
\title{Returns the bounding box of a \code{RhrTrack*}}
\usage{
rhrBBX(x, f, spatial, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}.}

\item{f}{Numeric value, fraction by which the bounding box is extended.}

\item{spatial}{Logical value, if \code{TRUE} a \code{SpatialPolygons} object is returned.}

\item{...}{None implemented.}
}
\value{
A matrix with the bounding box.
}
\description{
Returns the bounding box of a \code{RhrTrack*}
}
\examples{

data(trackS)
rhrBBX(trackS)

# Extends the range by 5\% to each side
rhrBBX(trackS, 0.05)


# Check that is actually works
bbx <- rhrBBX(trackS)
ext <- apply(bbx, 1, diff)

ext * 1.1
apply(rhrBBX(trackS, 0.05), 1, diff)  
}

