% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrBiNorm.R
\name{rhrBiNorm}
\alias{rhrBiNorm}
\title{rhrBiNorm}
\usage{
rhrBiNorm(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL), maxit = 20)
}
\arguments{
\item{xy}{valid input data}

\item{trast}{template raster}

\item{maxit}{Integer, giving the maximum number of iterations. See also \link{\code{mixtools::mvnormalmixEM}}, which is called.}
}
\description{
Bivariate bimodal Normal Home-Range estimation
}
\details{
Computes home-range using bivariate bimodal normal distribution
}

