% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrBurstify}
\alias{rhrBurstify}
\title{Burstify a track}
\usage{
rhrBurstify(x, minN = 3, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{minN}{Integer, the minimum number of relocations required for a birst.}

\item{...}{None implemented.}
}
\description{
Burstifying assumes, that the track has been regularized before. The smallest time interval is choosen as a reference interval.
}

