% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrCheckData.R
\name{rhrCheckData}
\alias{rhrCheckData}
\title{Check if valid data were provided}
\usage{
rhrCheckData(dat, returnSP = FALSE)
}
\arguments{
  \item{dat}{A data object.} \item{returnSP}{A logical value
  indicating whether or not a \code{SpatialPoints} should be
  returned.}
}
\value{
\code{data.frame}, \code{SpatialPoints} or \code{STI}
}
\description{
Check if the provided data meets certain criteria
}

