\name{rhrCoreArea.data.frame}
\alias{rhrCoreArea.data.frame}
\title{Core Area for data.frame}
\usage{
  \method{rhrCoreArea}{data.frame} (x, res, ...)
}
\arguments{
  \item{x}{a data.frame}

  \item{res}{the resolution of the raster}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Core Area for data.frame
}

