\name{rhrCoreArea.Raster}
\alias{rhrCoreArea.Raster}
\title{Core Area for Raster}
\usage{
  \method{rhrCoreArea}{Raster} (x, ...)
}
\arguments{
  \item{x}{Kernel density estimation}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Core Area for Raster
}

