\name{rhrCoreArea}
\alias{rhrCoreArea}
\title{Generic Function to calculate area of a RhrEstimator*}
\usage{
  rhrCoreArea(x, ...)
}
\arguments{
  \item{x}{a RhrEstimator* object}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Generic Function to calculate area of a RhrEstimator*
}

