\name{rhrCoreArea.RhrEstimatorKDE}
\alias{rhrCoreArea.RhrEstimatorKDE}
\title{Core Area for Kernel Density Estimation}
\usage{
  \method{rhrCoreArea}{RhrEstimatorKDE} (x, ...)
}
\arguments{
  \item{x}{Kernel density estimation}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Core Area for Kernel Density Estimation
}

