\name{rhrCoreArea.RhrEstimatorMCP}
\alias{rhrCoreArea.RhrEstimatorMCP}
\title{Core Area for Minimum Convex Polygon}
\usage{
  rhrCoreArea.RhrEstimatorMCP(est, ...)
}
\arguments{
  \item{est}{a RhrEstimatorMCP object}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Core Area for Minimum Convex Polygon
}

