\name{rhrCoreArea.RhrHREstimator}
\alias{rhrCoreArea.RhrHREstimator}
\title{Core area for home range estimator}
\usage{
  \method{rhrCoreArea}{RhrHREstimator} (x,
    method = "powell90", ...)
}
\arguments{
  \item{x}{object of class RhrHREstimator}

  \item{method}{method used to calculate core area}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  object of class RhrHRCoreArea
}
\description{
  Core area for home range estimator
}

