\name{rhrCorrectLevels}
\alias{rhrCorrectLevels}
\title{rhrCorrectLevels}
\usage{
  rhrCorrectLevels(str, low = 1, upper = 100,
    default.level = 95)
}
\arguments{
  \item{str}{A string with that contains the level,
  seperated by comas. If there are no suitable levels, then
  the default.level is used.}

  \item{low}{lower bound for levels}

  \item{upper}{upperbound for levels}

  \item{default.level}{is the default level}
}
\description{
  correct Levels at which a home range should be calculated
}

