% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrCRS}
\alias{rhrCRS}
\title{Coordinate Reference System (CRS) of a track}
\usage{
rhrCRS(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{...}{none implemented.}
}
\description{
Returns the proj4string.
}

