% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrCUD.R
\name{rhrCUD}
\alias{rhrCUD}
\title{Generic Function to calculate cumulative UD}
\usage{
rhrCUD(x, ...)
}
\arguments{
\item{x}{an object of class RhrHrEstimator*}

\item{...}{further arguments, none implemented}
}
\value{
an object of class raster
}
\description{
Generic Function to calculate cumulative UD
}

