\name{rhrEstimator}
\alias{rhrEstimator}
\title{Function to calculate a home range estimator}
\usage{
  rhrEstimator(xy, estimator, ...)
}
\arguments{
  \item{xy}{matrix or dataframe of dim n * 2, where n is
  the number of fixes}

  \item{estimator}{the estimator to be calculated}

  \item{...}{parameters passed to the home range estimator}
}
\value{
  object of class RhrEstimator
}
\description{
  Function to calculate a home range estimator
}

