\name{rhrEstimatorCharacteristic}
\alias{rhrEstimatorCharacteristic}
\title{Function to calculate characteristics of an estimator (i.e. area)}
\usage{
  rhrEstimatorCharacteristic(est, characteristic, ...)
}
\arguments{
  \item{est}{an object of class RhrEstimator*}

  \item{characteristic}{the name of the characteristic to
  be calculated}

  \item{...}{parameters passed to the characteristic
  function}
}
\value{
  object of class RhrEstimatorCharacteristic
}
\description{
  Function to calculate characteristics of an estimator
  (i.e. area)
}

