% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rhrExtFromPoints}
\alias{rhrExtFromPoints}
\title{rhrExtFromPoints}
\usage{
rhrExtFromPoints(xy, buffer = NULL, extendRange = NULL)
}
\arguments{
  \item{xy}{data.frame, matrix or SpatialPoints; containing
  the coordinates}

  \item{buffer}{numeric value; containing the buffer width}

  \item{extendRange}{numeric value; containing the factor
  to extend the range}
}
\value{
matrix
}
\description{
Takes coordinates from points and returns a (buffered,
extended) bounding box
}
\details{
The buffer width is the same in all directions. If both
buffer and extendRange are provided, precedence is given to
buffer.
}
\examples{
dat <- data.frame(x=rnorm(100), y=rnorm(100))
## Get the bounding box
rhrExtFromPoints(dat)
## Get the bounding box buffered by 10 units
rhrExtFromPoints(dat, buffer=10)
## Get the bounding box with an extended range
rhrExtFromPoints(dat, extendRange=0.5)
## Note, buffer takes precedence over extendRange
rhrExtFromPoints(dat, buffer=10, extendRange=0.5)
}
\author{
Johannes Signer
}
\seealso{
\code{\link{extendrange}} for more information on extending
the range.
}

