% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhrGUI.R
\name{rhrGUI}
\alias{rhrGUI}
\title{rhrGUI}
\usage{
rhrGUI(datFromR = NULL, outDir = getwd())
}
\arguments{
  \item{datFromR}{data.frame, that is taken from current
  namespace to the GUI}

  \item{outDir}{outdir}
}
\description{
Starts GUI
}
\details{
Start a web-based Graphical User Interface (GUI)
}

