% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rhrHasUD}
\alias{rhrHasUD}
\alias{rhrHasUD.RhrKDE}
\alias{rhrHasUD.RhrLoCoH}
\alias{rhrHasUD.RhrMCP}
\title{Generic Function to check if an ud is present}
\usage{
rhrHasUD(x, ...)

\method{rhrHasUD}{RhrMCP}(x, ...)

\method{rhrHasUD}{RhrKDE}(x, ...)

\method{rhrHasUD}{RhrLoCoH}(x, ...)
}
\arguments{
  \item{x}{an object of class RhrEst}

  \item{...}{further arguments, none implemented}
}
\value{
logical value indicating whether or not a UD is present
}
\description{
Generic Function to check if an ud is present
}

