% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrHrAnalysis.R
\name{rhrHrAnalysis}
\alias{rhrHrAnalysis}
\title{Function to run perform analyses and save results to temporary file}
\usage{
rhrHrAnalysis(dat, what = c("rhrSiteFidelity", "rhrTTSI", "rhrMCP", "rhrKDE",
  "rhrAsymptote", "rhrCoreArea"), args = NULL, verbose = TRUE,
  outDir = file.path(tempdir(), paste0("rhr", format(Sys.time(),
  "\%Y\%m\%d\%H\%M\%S"))), inGUI = FALSE, inUnit = "ido", outUnit = "ius",
  report = TRUE, zip = FALSE, repArgs = NULL)
}
\arguments{
\item{dat}{RhrMappedData}

\item{what}{character vector; indicating which steps should be performed, order does not matter}

\item{args}{list; with arguments for the different steps}

\item{verbose}{logical; whether or not progress should be given on the command line}

\item{outDir}{character; path to the directory where the results are saved.}

\item{inGUI}{logical; indicating whether or not the call was made from a GUI}

\item{inUnit}{character; units of the input}

\item{outUnit}{character; units of the output}

\item{report}{logical; indicating if a report should be created.}

\item{zip}{logical; indicating if a zip archive should be created, note this most likely only works under linux.}

\item{repArgs}{A list with extra arguments for the report.}
}
\value{
List
}
\description{
This functions wraps many rhr* methods. The purpose of this function to run a series of
of estimates given data, a vector of desired results and a list with arguments.
}
\author{
Johannes Signer
}

