% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrHrefScaled}
\alias{rhrHrefScaled}
\title{Select a Bandwidth for Kernel Density Estimation}
\usage{
rhrHrefScaled(xy, range = rhrHref(xy)$h * c(0.01, 2),
  trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange = 0.4), nrow =
  100, res = NULL), numOfParts = 1, level = 95, tol = 0.1, maxIt = 500)
}
\arguments{
  \item{xy}{A numeric data.frame with columns: x and y
  coordinates.}

  \item{range}{A numeric vector of length two, indicating
  the range of the search area. Be careful, if the search
  range is to large and trast to small, the algorithm will
  fail.}

  \item{trast}{A raster, to act as a template for kernel
  density estimates.}

  \item{numOfParts}{A numeric value indicating the number
  of polygons desired. This will usually be one.}

  \item{level}{A numeric value that specifies at which home
  range level the analysis should be performed.}

  \item{tol}{A numeric value indicating when the algorithm
  stops.}

  \item{maxIt}{A numeric value giving the maximum number of
  iterations acceptable.}
}
\value{
list with the found bandwidth and whether or not the
methods was successful.
}
\description{
Find bandwidth that results in n polygons (usually n=1) for
a Kernel Density Estimate.
}
\author{
Johannes Signer
}
\references{
Kie, John G. "A rule-based ad hoc method for selecting a
bandwidth in kernel home-range analyses." Animal
Biotelemetry 1.1 (2013): 1-12.
}

