\name{rhrHREstimator}
\alias{rhrHREstimator}
\title{rhrHREstimator}
\usage{
  rhrHREstimator(dat, call, params, ud, cud)
}
\arguments{
  \item{dat}{used data}

  \item{call}{how the fun was called}

  \item{params}{parameters that where passed}

  \item{ud}{whether or not an UD is desired}

  \item{cud}{whether or not a CUD is desired}
}
\description{
  Constructor for RhrHREstimator
}

