% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrInterpolate}
\alias{rhrInterpolate}
\title{Interpolate missing points}
\usage{
rhrInterpolate(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{...}{None implemented.}
}
\description{
Interpolating assumes, that the track has been regularized before. It picks the smallest time interval, an interpolates for all gaps that are larger than the minimum time interval. It is assumed, that all gaps are a multiplie of the smallest time gap.
}
\details{
Currently, rhrInterpolages loses all attributes and creates a new track.
}

