% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrIsopleths}
\alias{rhrIsopleths}
\title{rhrIsopleths}
\usage{
rhrIsopleths(x, ...)
}
\arguments{
  \item{x}{object of class RhrEst}

  \item{...}{see description}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Returns the isopleths of RhrEst object
}
\author{
Johannes Signer
}

