% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrLevels.R
\name{rhrLevels}
\alias{rhrLevels}
\title{Level(s) of Home Range Estimate}
\usage{
rhrLevels(x, ...)
}
\arguments{
\item{x}{Instance of \code{RhrEst}.}

\item{...}{none implemented.}
}
\value{
Numeric \code{vector} with the levels.
}
\description{
Function to retrieve the levels of a home range estimate.
}

