% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrMapFields}
\alias{rhrMapFields}
\title{to come}
\usage{
rhrMapFields(dat, fields = list(lon = NA, lat = NA, id = NA, date = NA, time =
  NA), proj4string = NULL, dateFormat = "ymd", timeFormat = "hms",
  defaultId = "Animal_1", duplicates)
}
\arguments{
  \item{dat}{}

  \item{fields}{}

  \item{proj4string}{}

  \item{dateFormat}{}

  \item{timeFormat}{}

  \item{defaultId}{}
}
\value{
RhrMappedData
}
\description{
to come
}
\details{
to come
}
\author{
Johannes Signer
}

