% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrMeta}
\alias{rhrMeta}
\title{Extract meta data from a track}
\usage{
rhrMeta(x, ...)
}
\arguments{
\item{x}{list.}

\item{...}{none implemented.}
}
\description{
Extract meta data from a track
}

