% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrN.R
\name{rhrN}
\alias{rhrN}
\title{Number of relocations in a trajectory}
\usage{
rhrN(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}} \item{missing}
  {logical scalar, should the number of missing locations be
  returned?}
}
\description{
Retunrs the number of relocations in a trajectory
}

