% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrPoints}
\alias{rhrPoints}
\title{Extract relocations from a track.}
\usage{
rhrPoints(x, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}

\item{...}{none implemented.}
}
\description{
Extract the relocation points from a track and return them as a `SpatialPointsPointsDataframe` object.
}

