% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrPrettyArea.R
\name{rhrPrettyArea}
\alias{rhrPrettyArea}
\title{Print home-range areas in a pretty way}
\usage{
rhrPrettyArea(x, inU, outU, ...)
}
\arguments{
\item{x}{An object of class RhrEst*.}

\item{inU}{Character scalar specifying input units.}

\item{outU}{Character scalar specifying output units.}

\item{...}{none implemented.}
}
\value{
A data.frame.
}
\description{
Creates a table of home-range areas and optionally converts units.
}
\examples{
library(rhr)
data(datSH)

## Geometric estimator
mcp <- rhrMCP(datSH[, 2:3])
rhrArgs(mcp)
}
\seealso{
\code{rhrConvertUnit}
}

