% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrPrettyArgs.R
\name{rhrPrettyArgs}
\alias{rhrPrettyArgs}
\title{Display Arguments of a Function in \code{rmd}}
\usage{
rhrPrettyArgs(x, as.rmd = TRUE, includeName = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{Rhr*}.}

\item{as.rmd}{Logical value, indicating if result should be coerced to \code{rmd}.}

\item{includeName}{Logical value indicating if the estimator name should be included or not.}
}
\description{
Function to return arguments of a \code{Rhr*}-Object as \code{data.frame} or \code{rmd} table.
}

