% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrSegments}
\alias{rhrSegments}
\title{Extract segments from a track.}
\usage{
rhrSegments(x, spatial, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}

\item{spatial}{Logical value, should the result be spatial (SpatialLinesDataFrame).}

\item{...}{none implemented.}
}
\description{
Extract the segments, linear interpolations between two points, from the track.
}

