\name{rhrSetCUD}
\alias{rhrSetCUD}
\title{Generic Function to set CUD}
\usage{
  rhrSetCUD(x, cud, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{cud}{the ud, object of class raster}

  \item{...}{further arguments, none implemented}
}
\value{
  an object of class rhrHREstimator with a cud
}
\description{
  Generic Function to set CUD
}

