\name{rhrSetCUD.RhrHREstimator}
\alias{rhrSetCUD.RhrHREstimator}
\title{Set the CUD}
\usage{
  \method{rhrSetCUD}{RhrHREstimator} (x, cud, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{cud}{the cud, object of class raster}

  \item{...}{further arguments, none implemented}
}
\description{
  Set the CUD
}

