\name{rhrSetIso.RhrHREstimator}
\alias{rhrSetIso.RhrHREstimator}
\title{Set isopleth}
\usage{
  \method{rhrSetIso}{RhrHREstimator} (x, iso, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{iso}{the isoplethes}

  \item{...}{further arguments, none implemented}
}
\description{
  Set isopleth
}

