\name{rhrSetUD}
\alias{rhrSetUD}
\title{Generic Function to set UD}
\usage{
  rhrSetUD(x, ud, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{ud}{the ud, object of class raster}

  \item{...}{further arguments, none implemented}
}
\value{
  an object of class rhrHREstimator with ud
}
\description{
  Generic Function to set UD
}

