\name{rhrSetUD.RhrHREstimator}
\alias{rhrSetUD.RhrHREstimator}
\title{Set the UD}
\usage{
  \method{rhrSetUD}{RhrHREstimator} (x, ud, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimator}

  \item{ud}{the ud, object of class raster}

  \item{...}{further arguments, none implemented}
}
\description{
  Set the UD
}

