% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrSplit}
\alias{rhrSplit}
\title{Split a track}
\usage{
rhrSplit(x, f, minN, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{f}{Vector indicating the grouping}

\item{minN}{Integer, the minimum number of relocations required in a group in order to create a new track.}

\item{...}{none implemented.}
}
\description{
Returns a list of subtracks.
}

