% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrTimes}
\alias{rhrTimes}
\title{Extract time stamps from a track.}
\usage{
rhrTimes(x, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}
}
\description{
Extracts the time stamps from a track and return them as a vector.
}
\examples{
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- lubridate::ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackST (only space)
trackST <- rhrTrack(sp, time)
time <- rhrTimes(trackST)
}

