% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrack.R
\name{rhrTrack}
\alias{rhrTrack}
\alias{rhrTracks}
\title{Create objects of class `RhrTrack` to represent animal relocation data.}
\usage{
rhrTrack(sp, time, duplicates = "remove", meta)

rhrTracks(sp, ts, id, metas)
}
\arguments{
\item{sp}{SpatialPoints* with the relocation of the animal or \code{RhrMappedData}.}

\item{time}{Vector with objects of class `POSIXct`, giving the time stamp of relocations.}

\item{duplicates}{String, indicating how to handle duplicted time stamps. Currently only  "remove" is supported.}

\item{meta}{Named list, providing meta information about the animal.}

\item{ts}{Vector containing the timestamp for each point in \code{sp}.}

\item{id}{Vector containing the id for each point in \code{sp}.}

\item{metas}{List of list with meta data, length of \code{metas} should be the same as the number of levels in \code{id}.}
}
\description{
Creates object of class `RhrTrack*` from `SpatialPoints*` and, if provided, time stamps.
}

