% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrTrackData}
\alias{rhrTrackData}
\title{Returns data of a RhrTrack* object.}
\usage{
rhrTrackData(x, ...)
}
\arguments{
\item{x}{RhrTrack* object.}

\item{...}{ignored}
}
\description{
Returns data of a RhrTrack* object.
}

