% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajStats.R
\name{rhrTrackStats}
\alias{rhrTrackStats}
\title{Retunrs trajectory statistics}
\usage{
rhrTrackStats(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}}
}
\value{
Object of class \code{data.frame}
}
\description{
Retunrs trajectory statistics
}

