% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTraj.R
\name{rhrTraj}
\alias{rhrTraj}
\title{Constructur for a Trajectory}
\usage{
rhrTraj(sp, time = NULL)
}
\arguments{
  \item{sp}{Object of class SpatialPoints, the relocation.
  If `sp` is an object of `SpatialPointsDataFrame`, point
  attributes are carried forward.} \item{time}{Object
  of class `POSIXct`, the time when the relocations were
  recorded. If `time` is NULL without time stamps is
  created.}
}
\value{
Object of class \code{RhrTrajS*}
}
\description{
Creates an object of `RhrTraj*` from `SpatialPoints`.
}
\examples{
library(rhr)

data(datSH)

ts <- lubridate::ymd(datSH$day) + lubridate::hms(datSH$time)
pts <- sp::SpatialPoints(datSH[, 2:3])

traj <- rhrTraj(pts, ts)
}

