% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajAttribute.R
\name{rhrTrajAttribute}
\alias{rhrTrajAttribute}
\title{Retunrs trajectory attributes}
\usage{
rhrTrajAttribute(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}}
}
\value{
Object of class \code{data.frame}
}
\description{
Retunrs trajectory attributes
}

