% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajBurstify.R
\name{rhrTrajBurstify}
\alias{rhrTrajBurstify}
\title{Splits a trajectory into bursts.}
\usage{
rhrTrajBurstify(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajSTR}}
}
\description{
Splits a regular trajectory into burts (i.e., sequence of
relocations with no missing spatial information).
}

