% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajGaps.R
\name{rhrTrajGaps}
\alias{rhrTrajGaps}
\title{Gaps of a trajectory}
\usage{
rhrTrajGaps(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajSTR}}
}
\description{
Extracts gaps of a regular trajectory (i.e., time-stamps
with missing spatial information).
}

