% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajLocations.R
\name{rhrTrajLocations}
\alias{rhrTrajLocations}
\title{Retrieves the locations of a  trajectory}
\usage{
rhrTrajLocations(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}}
}
\value{
Object of class \code{SpatialPoints} or \code{STIDF}
}
\description{
Retruns the location of a trajectory as
\code{sp::SpatialPoints} for trajectories without temporal
information and as \code{spacetime::STIDF} for trajectories
with temporal information.
}

