% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajRelocations.R
\name{rhrTrajRelocations}
\alias{rhrTrajRelocations}
\title{Retrieve relocations of a trajectory}
\usage{
rhrTrajRelocations(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}}
}
\value{
Object of class \code{SpatialPoints}
}
\description{
Retruns the relocation of a trajectory as
\code{SpatialPoints}.
}

