% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajSplit.R
\name{rhrTrajSplit}
\alias{rhrTrajSplit}
\title{Split trajectory}
\usage{
rhrTrajSplit(x, f, ...)
}
\arguments{
  \item{x}{RhrTraj* object.} \item{f}{factor that provides a
  grouping for splitting data.}
}
\description{
Split a trajectory into a list of new trajectories.
}

