% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrajTimes.R
\name{rhrTrajTimes}
\alias{rhrTrajTimes}
\title{Time of relocations}
\usage{
rhrTrajTimes(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}} \item{missing}
  {logical scalar, should the number of missing locations be
  returned?}
}
\description{
Retunrs the time a relocation was recorded
}

