% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTuningParameter.R
\name{rhrTuningParameter}
\alias{rhrTuningParameter}
\title{Tuning parameter of Home Range Estimate}
\usage{
rhrTuningParameter(x, msg = FALSE, digits = 3, ...)
}
\arguments{
  \item{x}{Instance of \code{RhrEst}.} \item{msg}{Boolean
  scalar, indicating if the answer should be returned
  as message.} \item{digits}{Numeric scalar, giving the
  number of decimal places to round to.} \item{...}{none
  implemented}
}
\value{
If \code{msg} is \code{TRUE} a character string, else a list
with the name of the tuning parameter and its value.
}
\description{
Function to retrieve the tuning parameter(s) of a home range
estimate.
}
\details{
Some home range estimators require a tuning parameter
for the estimation (e.g. the bandwidth for kernel density
estimates). This functions retrieves the value of the tuning
parameter. If \code{msg = TRUE} values are rounded to the
\code{digit} decimal place.
}
\examples{
library(rhr)
data(datSH)

kde <- rhrKDE(datSH[, 2:3])
locoh <- rhrLoCoH(datSH[, 2:3])

rhrTuningParameter(kde)
rhrTuningParameter(kde, msg = TRUE)
rhrTuningParameter(kde, msg = TRUE, digits = 1)
rhrTuningParameter(kde, msg = TRUE, digits = 10)

rhrTuningParameter(locoh)
rhrTuningParameter(locoh, msg = TRUE)
}

