% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrUD}
\alias{rhrUD}
\title{Utilization distribution}
\usage{
rhrUD(x, ...)
}
\arguments{
  \item{x}{An object of class RhrProbEst.}

  \item{...}{none implemented.}
}
\value{
An object of class RasterLayer.
}
\description{
Function to obtain the utilization distribution (UD) of a
home range estimate.
}
\details{
For probabilistic estimators home range estimator this
functions returns the UD.
}
\examples{
library(rhr)
data(datSH)

## For KDE
kde <- rhrKDE(datSH[, 2:3])
ud <- rhrUD(kde)
plot(rhrUD(kde))


## For unimodal normal
un <- rhrUniNorm(datSH[, 2:3])
ud <- rhrUD(un)
plot(ud)
}

