% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrUniNorm}
\alias{rhrUniNorm}
\title{rhrUniNorm}
\usage{
rhrUniNorm(xy, trast = NULL, proj4string = NA)
}
\arguments{
  \item{xy}{valid input data}

  \item{trast}{template raster}

  \item{proj4string}{proj4 string}
}
\description{
Univariate Normal Home-Range estimation
}
\details{
Computes home-range using univariate normal distribution
}

