% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrValidEpsg.R
\name{rhrValidEpsg}
\alias{rhrValidEpsg}
\title{rhrValidEpsg}
\usage{
rhrValidEpsg(epsg)
}
\arguments{
\item{epsg}{EPSG code}
}
\description{
Validate EPSG code
}
\details{
Takes an EPSG checks if it is valid
}

