% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrWithin}
\alias{rhrWithin}
\title{Spatial subset of a track}
\usage{
rhrWithin(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}.}

\item{y}{Object of class \code{SpatialPolygons*}}

\item{...}{none implemented.}
}
\value{
Object of class \code{rhrTrack*}.
}
\description{
Performs a subset of a track based on the spatial position of relocations. Only relocations that are within a polygon (\code{SpatialPolygons*}) are selected a new track is created.
}

