% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrWithinTime}
\alias{rhrWithinTime}
\title{Temporal subset of a track}
\usage{
rhrWithinTime(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrackST*} or code \code{RhrTracksST}.}

\item{y}{Object of class \code{lubridate::Interval}}

\item{...}{None implemented.}
}
\value{
Object of class \code{rhrTrack*}.
}
\description{
Performs a subset of a track based on a provided time interval. Only relocations that are within the time interval are selected and (a) new track(s) is created.
}

