\name{strShorten}
\alias{strShorten}
\title{strShorten}
\usage{
  strShorten(str, max.nchar = 20, cont = "...")
}
\arguments{
  \item{str}{a string}

  \item{max.nchar}{maximum number of chars in the return
  string}

  \item{cont}{ending of the shortend string}
}
\description{
  Shortens a string
}

