\name{alert}
\alias{alert}
\title{alert}
\usage{
  alert(x, cat = TRUE)
}
\arguments{
  \item{x}{a string}

  \item{cat}{if this \code{cat=TRUE} then the return string
  is wrapped with a \code{cat}.}
}
\description{
  wraps a string with html markup for a twitter bootstrap
  alert
}
\examples{
alert("Calculation completed")
}
\author{
  Johannes Signer
}

