% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{datSH}
\alias{datSH}
\title{Tracking data of one red deer in Germany}
\format{1 location per row}
\source{
Verein fuer Wildtierforschung Goettingen and Dresden
}
\usage{
datSH
}
\description{
data set from one red deer from northern Germany
}

