\name{dfGrob}
\alias{dfGrob}
\title{dfGrob}
\usage{
  dfGrob(x, start = 0.03, stop = 0.97, digits = 2)
}
\arguments{
  \item{x}{a data.frame}

  \item{start}{where to start, default is 0.03}

  \item{stop}{where to stop, default is 0.97}

  \item{digits}{to how many digists numbers are rounded}
}
\description{
  Create a grid representation for an object of class
  data.frame
}

