% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{is.regular}
\alias{is.regular}
\title{Checks if the track is regularly spaced in time.}
\usage{
is.regular(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrackS*}}

\item{...}{none implemented.}
}
\description{
Returns `TRUE` if time differences between consecutive relocations are equal.
}

