% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTraj.R
\name{is.regular}
\alias{is.regular}
\title{Method to check if a trajectory is regularely spaced in time.}
\usage{
is.regular(x, ...)
}
\arguments{
  \item{x}{Object of \code{RhrTrajS*}.} \item{...}{additional
  arguments, none implemented.}
}
\description{
Method to check if a trajectory is regularely spaced in
time.
}

