% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrBiCir.R, R/rhrParamEst-Methods.R, R/rhrUniCir.R
\name{rhrBiCirc}
\alias{parametric_homeranges}
\alias{rhrBiCirc}
\alias{rhrUniCirc}
\title{rhrBiCirc}
\usage{
rhrBiCirc(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL))

rhrUniCirc(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL))
}
\arguments{
\item{xy}{Instance of \code{data.frame}, \code{SpatialPoints},
\code{SpatialPointsDataFrame} or \code{RhrMappedData}. If \code{xy} is
is a \code{data.frame}, then the first two columns are assumed to be the
\code{x} and \code{y} coordinates of the relocations.}

\item{trast}{Instance of \code{RasterLayer} that serves as a template for the resulting grid of the estimation.}
}
\description{
Bimodal circular home range estimation
}
\details{
Computes home-range using two circular normal distributions
}

