% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.RhrHRAsymptote}
\alias{plot.RhrHRAsymptote}
\title{plot for RhrHRAsymptote}
\usage{
\method{plot}{RhrHRAsymptote}(x, ...)
}
\arguments{
  \item{x}{RhrHRAsymptote}

  \item{...}{none implemented}
}
\description{
generic plot for RhrHRAsymptote
}

