\name{plot.RhrHRAsymptote}
\alias{plot.RhrHRAsymptote}
\title{plot for RhrHRAsymptote}
\usage{
  \method{plot}{RhrHRAsymptote} (x, draw = TRUE, ...)
}
\arguments{
  \item{x}{RhrHRAsymptote}

  \item{draw}{indicates whether the plot should be drawn or
  not. If this is \code{FALSE} a grob is returned.}

  \item{...}{none implemented}
}
\description{
  generic plot for RhrHRAsymptote
}

