% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{plot.RhrTrack}
\alias{plot.RhrTrack}
\title{Prints RhrTrack object.}
\usage{
\method{plot}{RhrTrack}(x, ...)
}
\arguments{
\item{x}{RhrTrack object.}

\item{...}{None implemented.}
}
\description{
Prints RhrTrack object.
}

