% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{plot.RhrTracks}
\alias{plot.RhrTracks}
\title{Prints RhrTracks object.}
\usage{
plot.RhrTracks(x, ...)
}
\arguments{
\item{x}{RhrTracks object.}

\item{...}{None implemented.}
}
\description{
Prints RhrTracks object.
}

