% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.RhrTTSI}
\alias{plot.RhrTTSI}
\title{plot for RhrTTSI
generic plot for RhrTTSI}
\usage{
\method{plot}{RhrTTSI}(x, ...)
}
\arguments{
  \item{x}{RhrHREstimator object}

  \item{...}{none implemented}
}
\description{
plot for RhrTTSI generic plot for RhrTTSI
}

