\name{plotTTSI}
\alias{plotTTSI}
\title{plot ttsi}
\usage{
  plotTTSI(res, path = NULL, toFirst2 = FALSE)
}
\arguments{
  \item{res}{from ttsi}

  \item{path}{where the plot should be safed}

  \item{toFirst2}{only plot until critical value of 2 is
  first reached}

  \item{returnGrob}{retun a grob}
}
\description{
  plots results of rhrSchoener
}

