% Generated by roxygen2 (4.0.2): do not edit by hand
\name{progressInit}
\alias{progressInit}
\title{Initialize progress
This function is copied from shinyIncubator and will be removed as soon as the progress function moved to shiny}
\usage{
progressInit()
}
\description{
Call this function in your \code{shinyUI} definition if you
intend to use progress in \code{server.R}.
}
\seealso{
\code{\link{withProgress}}, \code{\link{Progress}}
}

