% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrRasterFromExtent.R
\name{rasterFromXYVect}
\alias{rasterFromXYVect}
\title{Create \code{RasterLayer} from the extent of points.}
\usage{
rasterFromXYVect(xy, xrange = NA, yrange = NA, res = 100)
}
\arguments{
  \item{xy}{A two columned \code{data.frame} or
  \code{matrix} containing the x and y coordinates of the
  points.} \item{xrange,yrange}{Optional numeric vector of
  length providing the range.} \item{res}{resolution}
}
\description{
Function to create an empty \code{RasterLayer} from bounding
box of points defined by x and y coordinates.
}
\details{
Takes x and y coordinates of points as input and creates
a \code{RasterLayer} with the input coordiantes. The range
in either direction can be provided, otherwise it will be
obtained from the data.
}
\examples{
library(rhr)
data(datSH)
(r <- rasterFromXYVect(datSH[, 2:3]))
}

