% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrArea}
\alias{rhrArea}
\title{rhrArea}
\usage{
rhrArea(x, ...)
}
\arguments{
  \item{x}{object of class RhrEst}

  \item{...}{none implemented}
}
\value{
vector
}
\description{
Returns the area of RhrEst object
}
\author{
Johannes Signer
}

