% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rhrArea.R
\name{rhrArea}
\alias{rhrArea}
\title{Area of Home Range Estimate}
\usage{
rhrArea(x, ...)
}
\arguments{
\item{x}{Instance of \code{RhrEst}.}

\item{...}{see details.}
}
\value{
\code{data.frame} with the isopleth level and area in units of the coordinate system.
}
\description{
Function to calculate the area of a home range estimate.
}
\details{
For geometric estimators the isopleth levels at which the area is calculated is
already determined when the estimate is performed. Probabilistic estimators take
an additional argument, \code{levels}, that determines at which isopleth levels
of the UD the area is evaluated.
}
\examples{
library(rhr)
data(datSH)

## Geometric estimator
mcp <- rhrMCP(datSH[, 2:3])
rhrArea(mcp)

mcp <- rhrMCP(datSH[, 2:3], levels = seq(50, 95, 5))
rhrArea(mcp)

## Probabilistic estimator
kde <- rhrKDE(datSH[, 2:3])
rhrArea(kde)

## at a different isopleth level
rhrArea(kde, levels = 50)

## or at sequence of isopleth levels
rhrArea(kde, level = seq(50, 95, 5))
}

