\name{rhrArea}
\alias{rhrArea}
\title{Generic Function to calculate area of a RhrEstimator*}
\usage{
  rhrArea(est, ...)
}
\arguments{
  \item{est}{a RhrEstimator* object}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Generic Function to calculate area of a RhrEstimator*
}

