% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrBBMM}
\alias{rhrBBMM}
\title{Brownian Bridge Movement Model (BBMM)}
\usage{
rhrBBMM(xy, time, rangesigma1 = c(0, 10000), sigma2 = 123,
  trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange = 0.2), nrow =
  100, res = NULL))
}
\arguments{
  \item{xy}{\code{data.frame} with two columns: x and y
  coordinates.}

  \item{time}{a vector with time stamps for each
  relocation.}

  \item{rangesigma1}{parameter 1}

  \item{sigma2}{parameter 2}

  \item{trast}{a \code{RasterLayer} used as an template for
  the output grid.}
}
\value{
object of class \code{RhrBBMM}
}
\description{
A function to estimate home ranges with kernel density
estimation.
}
\author{
Johannes Signer
}
\seealso{
\code{adehabitatHR::kernelbb}
}

