% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rhrBBMM}
\alias{rhrBBMM}
\title{Brownian Bridge Movement Model (BBMM)}
\usage{
rhrBBMM(xyt, rangesigma1 = c(0, 10000), sigma2 = 123,
  trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange = 0.2), nrow =
  100, res = NULL), proj4string = NA)
}
\arguments{
  \item{xy}{\code{data.frame} with two columns: x and y
  coordinates.}

  \item{sigma1}{}

  \item{sigma2}{}

  \item{trast}{a \code{RasterLayer} used as an template for
  the output grid.}
}
\value{
object of class \code{RhrBBMM}
}
\description{
A function to estimate home ranges with kernel density
estimation.
}
\author{
Johannes Signer
}
\seealso{
\code{KernSmooth::bkde2d}, \code{KernSmooth::dpik},
\code{rhr::rhrHref}, \code{rhr::rhrHlscv},
\code{rhr::rhrHpi}
}

