% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrBBMM.R
\name{rhrBBMM}
\alias{rhrBBMM}
\title{Brownian Bridge Movement Model (BBMM)}
\usage{
rhrBBMM(xy, time, rangesigma1 = c(0, 10000), sigma2 = 100,
  trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange = 0.2), nrow =
  100, res = NULL))
}
\arguments{
  \item{xy}{\code{Data.frame} with two columns containing
  x and y coordinates.} \item{time}{a vector with time
  stamps for each relocation, must be of the same length as
  \code{xy}.} \item{rangesigma1}{parameter 1} \item{sigma2}
  {parameter 2} \item{trast}{a \code{RasterLayer} used as an
  template for the output grid.}
}
\value{
object of class \code{RhrBBMM}
}
\description{
A wrapper around \code{adehabitatHR::kernelbb}.
}
\seealso{
\code{adehabitatHR::kernelbb}
}

