% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrBBX}
\alias{rhrBBX}
\title{Returns the bounding box of a \code{RhrTrack*}}
\usage{
rhrBBX(x, f, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}.}

\item{f}{Numeric value, fraction by which the bounding box is extended.}

\item{...}{none implemented.}
}
\value{
A matrix with the bounding box.
}
\description{
Returns the bounding box of a \code{RhrTrack*}
}
\examples{
data(trackS)

rhrBBX(trackS)

# Extends the rangy by 5\% to each side
rhrBBX(trackS, 0.05)


# Check that is actually works
bbx <- rhrBBX(trackS)
ext <- apply(bbx, 1, diff)

ext * 1.1
apply(rhrBBX(trackS, 0.05), 1, diff)  
}

