% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrBBX}
\alias{rhrBBX}
\title{Returns the bounding box of a \code{RhrTrack*}}
\usage{
rhrBBX(x, f, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}.}

\item{f}{Numeric value, fraction by which the bounding box is extended.}

\item{...}{none implemented.}
}
\value{
A matrix with the bounding box.
}
\description{
Returns the bounding box of a \code{RhrTrack*}
}

