% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrBiCirc}
\alias{rhrBiCirc}
\title{rhrBiCirc}
\usage{
rhrBiCirc(xy)
}
\arguments{
  \item{xy}{valid input data}
}
\description{
Bimodal circular home range estimation
}
\details{
Computes home-range using two circular normal distributions
}

