% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrBiNorm}
\alias{rhrBiNorm}
\title{rhrBiNorm}
\usage{
rhrBiNorm(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL))
}
\arguments{
  \item{xy}{valid input data}

  \item{trast}{template raster}
}
\description{
Bivariate bimodal Normal Home-Range estimation
}
\details{
Computes home-range using bivariate bimodal normal
distribution
}

