\name{rhrCoreArea.RhrEstimatorKDE}
\alias{rhrCoreArea.RhrEstimatorKDE}
\title{Core Area for Kernel Density Estimation}
\usage{
  rhrCoreArea.RhrEstimatorKDE(est, ...)
}
\arguments{
  \item{est}{Kernel density estimation}
}
\value{
  The area
}
\description{
  Core Area for Kernel Density Estimation
}

