\name{rhrCoreArea.RhrEstimatorMCP}
\alias{rhrCoreArea.RhrEstimatorMCP}
\title{Core Area for Minimum Convex Polygon}
\usage{
  \method{rhrCoreArea}{RhrEstimatorMCP} (x, ...)
}
\arguments{
  \item{x}{a RhrEstimatorMCP object}

  \item{...}{further parameters passed to the area
  function}
}
\value{
  The area
}
\description{
  Core Area for Minimum Convex Polygon
}

