% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrCorrectLevels}
\alias{rhrCorrectLevels}
\title{rhrCorrectLevels}
\usage{
rhrCorrectLevels(str, low = 1, upper = 100, defaultLevel = 95)
}
\arguments{
  \item{str}{A string that contains the level, seperated by
  comas (usually provided through the GUI). If there are no
  suitable levels, then the default.level is used.}

  \item{low}{lower bound for levels}

  \item{upper}{upperbound for levels}

  \item{defaultLevel}{is the default level}
}
\description{
correct levels at which a home range should be calculated
}

