% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrCUD}
\alias{rhrCUD}
\title{Generic Function to calculate cumulative UD}
\usage{
rhrCUD(x, ...)
}
\arguments{
  \item{x}{an object of class RhrHrEstimator*}

  \item{...}{further arguments, none implemented}
}
\value{
an object of class raster
}
\description{
Generic Function to calculate cumulative UD
}

