% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhrKDE.R
\name{rhrCUD.RhrKDE}
\alias{rhrCUD.RhrKDE}
\title{Calculate cumulative UD for RhrHrEstimatorKDE}
\usage{
\method{rhrCUD}{RhrKDE}(x, ...)
}
\arguments{
  \item{x}{an object of class rhrHREstimatorKDE}

  \item{...}{further arguments, none implemented}
}
\description{
Calculate cumulative UD for RhrHrEstimatorKDE
}

