% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrData.R
\name{rhrData}
\alias{rhrData}
\title{Return Data}
\usage{
rhrData(x, spatial = FALSE, ...)
}
\arguments{
  \item{x}{Instance of \code{RhrEst}.} \item{spatial}
  {Logical value indicating whether or not to return
  \code{SpatialPoints}.} \item{...}{none implemented.}
}
\value{
Depending on the argument, either a \code{data.frame} or
\code{SpatialPointsDataFrame}.
}
\description{
Function to retrieve data that were used generate the
estimate.
}
\examples{
library(rhr)
data(datSH)

head(rhrData(rhrMCP(datSH[, 2:3])))
str(rhrData(rhrMCP(datSH[, 2:3]), spatial = TRUE))
}

