% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrData}
\alias{rhrData}
\title{rhrData}
\usage{
rhrData(x, spatial = FALSE, ...)
}
\arguments{
  \item{x}{object of class RhrEst}

  \item{spatial}{logical value indicating whether or not to
  return spatial points}

  \item{...}{}
}
\value{
data.frame
}
\description{
Returns the area of RhrEst object
}
\author{
Johannes Signer
}

