% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrHasTS.R
\name{rhrHasTS}
\alias{rhrHasTS}
\title{Checks if trajectory has a time stamp}
\usage{
rhrHasTS(x, ...)
}
\arguments{
  \item{x}{Object of class \code{RhrTrajS*}}
}
\description{
Retunrs TRUE if the trajectory has a time stamp and FALSE if
not.
}

