% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrHasTS}
\alias{rhrHasTS}
\title{Checks if the track has time stamps}
\usage{
rhrHasTS(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{...}{none implemented.}
}
\description{
Retunrs \code{TRUE} if the track has time stamps.
}
\examples{
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- lubridate::ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackS (only space)
trackS <- rhrTrack(sp)

## Create an object of RhrTrackST (only space)
trackST <- rhrTrack(sp, time)

rhrHasTS(trackS)
rhrHasTS(trackST)
}

