% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrHasTS}
\alias{rhrHasTS}
\title{Checks if the track has time stamps}
\usage{
rhrHasTS(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{...}{none implemented.}
}
\description{
Retunrs \code{TRUE} if the track has time stamps.
}

