% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrHasUD.RhrBBMM}
\alias{rhrHasUD}
\alias{rhrHasUD.RhrBBMM}
\alias{rhrHasUD.RhrBiCirc}
\alias{rhrHasUD.RhrBiNorm}
\alias{rhrHasUD.RhrGeoEst}
\alias{rhrHasUD.RhrKDE}
\alias{rhrHasUD.RhrUniNorm}
\title{Generic Function to check if an ud is present}
\usage{
\method{rhrHasUD}{RhrBBMM}(x, ...)

\method{rhrHasUD}{RhrBiCirc}(x, ...)

\method{rhrHasUD}{RhrBiNorm}(x, ...)

rhrHasUD(x, ...)

\method{rhrHasUD}{RhrGeoEst}(x, ...)

\method{rhrHasUD}{RhrKDE}(x, ...)

\method{rhrHasUD}{RhrUniNorm}(x, ...)
}
\arguments{
  \item{x}{an object of class RhrEst}

  \item{...}{further arguments, none implemented}
}
\value{
logical value indicating whether or not a UD is present
}
\description{
Generic Function to check if an ud is present
}

