% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrHasUD.R, R/rhrKDE.R
\name{rhrHasUD}
\alias{rhrHasUD}
\alias{rhrHasUD.RhrGeoEst}
\alias{rhrHasUD.RhrKDE}
\title{Generic Function to check if an ud is present}
\usage{
rhrHasUD(x, ...)

\method{rhrHasUD}{RhrGeoEst}(x, ...)

\method{rhrHasUD}{RhrKDE}(x, ...)
}
\arguments{
\item{x}{an object of class RhrEst}

\item{...}{further arguments, none implemented}
}
\value{
logical value indicating whether or not a UD is present
}
\description{
Generic Function to check if an ud is present
}

