% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrHCV.R
\name{rhrHcv}
\alias{rhrHcv}
\title{Estimate bandwidth with likelihood cross validation}
\usage{
rhrHcv(xy)
}
\arguments{
\item{xy}{data.frame with two columns: x and y coordinates}
}
\value{
\code{vector} of length two
}
\description{
Likelihood Cross Validation
}
\examples{
data(datSH)
rhrHpi(datSH[, 2:3])
}
\author{
Johannes Signer
}

