% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrHCV.R
\name{rhrHcv}
\alias{rhrHcv}
\title{Estimate bandwidth with likelihood cross validation}
\usage{
rhrHcv(xy)
}
\arguments{
\item{xy}{data.frame with two columns: x and y coordinates}
}
\value{
\code{vector} of length two
}
\description{
Likelihood Cross Validation
}
\author{
Johannes Signer 
data(datSH)
rhrHpi(datSH[, 2:3])
}

