% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrHpi}
\alias{rhrHpi}
\title{Calculate a Bandwidth for Kernel Density Estimation.}
\usage{
rhrHpi(xy, rescale = "none", correct = TRUE, ...)
}
\arguments{
  \item{xy}{data.frame with two columns: x and y
  coordinates}

  \item{rescale}{character, indicating if and if how data
  are rescaled. \code{unitvar} rescales x and y coordinates
  to unit variance, \code{xvar} rescales x and y coordinate
  to variance of x and \code{none} uses the raw data.}

  \item{correct}{logical value that indicates whether or
  not the estimate should be correct for the two
  dimensional case, see details.}

  \item{...}{additional arguments passed to
  \code{KernSmooth::dpik}.}
}
\value{
\code{vector} of length two
}
\description{
Wraps \code{KernSmooth::dpik} to estimate bandwidth for
kernel density estimation in the two dimensional case.
}
\details{
This function calcualtes bandwidth values for kernel
density estimation by wrapping \code{KernSmooth::dpik}. If
\code{correct} is TRUE, the bandwidth is trasformed with
power 5/6 to correct for using an univariate implementation
for bivariate data (Gitzen et. al 2006).
}
\examples{
data(datSH)
rhrHpi(datSH[, 2:3])
}
\author{
Johannes Signer
}
\references{
Gitzen, R. A., Millspaugh, J. J., & Kernohan, B. J. (2006).
Bandwidth selection for fixed-kernel analysis of animal
utilization distributions. _Journal of Wildlife
Management_, 70(5), 1334-1344.
}
\seealso{
\code{KernSmooth::dpik}
}

