% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrHrAnalysis}
\alias{rhrHrAnalysis}
\title{Function to run perform analyses and save results to temporary file}
\usage{
rhrHrAnalysis(datIn, what = c("rhrSiteFidelity", "rhrTTSI", "rhrMCP",
  "rhrKDE", "rhrAsymptote", "rhrCoreArea"), args = NULL, verbose = TRUE,
  outDir = file.path(tempdir(), paste0("rhr", format(Sys.time(),
  "\%Y\%m\%d\%H\%M\%S"))), inGUI = FALSE, inUnit = "ido", outUnit = "ius",
  writeResultsSpatial = TRUE)
}
\arguments{
  \item{what}{character vector; indicating which steps
  should be performed, order does not matter}

  \item{args}{list; with arguments for the different steps}

  \item{verbose}{logical; whether or not progress should be
  given on the command line}

  \item{dat}{RhrMappedData}

  \item{duplicated}{character; determine what to do with
  either 'nothing', 'delete' or 'random', to add random
  noise.}
}
\value{
List
}
\description{
This functions wraps many rhr* methods. The purpose of this
function to run a series of of estimates given data, a
vector of desired results and a list with arguments.
}
\author{
Johannes Signer
}

