% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrKDE}
\alias{rhrKDE}
\title{Kernel Density Estimation (KDE)}
\usage{
rhrKDE(xy, h = rhrHref(xy)$h, levels = 95,
  trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange = 0.2), nrow =
  100, res = NULL))
}
\arguments{
  \item{xy}{\code{data.frame} with two columns: x and y
  coordinates.}

  \item{h}{character ("href" or "hlscv") specifying the
  method to estimate the bandwidth or numeric value
  specifying the bandwidth.}

  \item{levels}{A numeric vector, indicating isopleths of
  KDE.}

  \item{trast}{a \code{RasterLayer} used as an template for
  the output grid.}
}
\value{
object of class \code{RhrHREstimator}
}
\description{
A function to estimate home ranges with kernel density
estimation.
}
\details{
Kernels densities are estimated with
\code{KernSmooth::bkde2d}. This is a binned approximation
of 2D kernel density estimates (see
\code{?KernSmooth::bkde2d} for more details.
}
\examples{
data(datSH)
\dontrun{
# Kernel with href bandwidth estimation
k1 <- rhrKDE(datSH[, 2:3], h="href", res=100)
plot(k1)

# what is the actually estimated bandwidth?
k1$parameters$h

# Kernel with href bandwidth estimation
k2 <- rhrKDE(datSH[, 2:3], h="lscv", res=100)
plot(k2)

# what is the actually estimated bandwidth?
k2$parameters$h
}
}
\seealso{
\code{KernSmooth::bkde2d}, \code{rhr::rhrHref},
\code{rhr::rhrHlscv}, \code{rhr::rhrHpi}
}

