% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrLoCoH}
\alias{rhrLoCoH}
\title{Local Convex Hull (LoCoH)}
\usage{
rhrLoCoH(xy, type = "k", n = 10, levels = 95, minPts = 3,
  autoN = FALSE)
}
\arguments{
  \item{xy}{data.frame with two columns. The first column
  contains x coordinates and the second column contains y
  coordinates.}

  \item{type}{character value, one of k, r, a. Three
  methods to select which neighbours are used for local
  hulls are availabe: i) k-nearest neighbours; ii)
  neighbours within a radius r; iii) neighbours within a
  cummulative dististance a.}

  \item{n}{numeric value, if type is k it is the number of
  neaerst neibhers, if type is r it is the radius that is
  searched and if type is a it is the cummulative distance
  to be used.}

  \item{levels}{numeric vector, indicating desired
  isopleths.}

  \item{minPts}{numeric value, the minimum number of
  neighbours required.}

  \item{autoN}{boolean, determine \code{n} automatically.
  This take precedence over n, but is set to \code{FALSE}
  by default.}
}
\value{
object of class \code{RhrHREstimator}
}
\description{
This function estimates an animals home range using Local
Convex Hulls (LoCoH).
}
\details{
Three different types available for determining the number
of neighbors: \itemize{ \item{"k"}{uses the k nearest
neighbours} \item{"r"}{uses all neighbours within a radius
r} \item{"a"}{uses all neighbours that can be reached
within a distance a. The distance to all points is
calculated and then the cummulatively summed starting from
the smallest until \code{a} is reached.}}. \code{autoN} is
the square root of all locations, if \code{type} is "k". If
\code{type} is "a" then "n" will be the maximum distance
between any two relocations. If \code{type} is "r" the
distance of "n" will be 5% of the maximum distance between
any two points.
}
\examples{
\dontrun{
data(datSH)
locoh <- rhrLoCoH(datSH[, 2:3], type="k", n=10)

## area at isopleths
rhrArea(locoh)

## get isopleths
iso <- isopleths(locoh)

## Which parameter were used?
parameters(locoh)
}
}
\author{
Johannes Signer
}
\references{
Getz, W. M., & Wilmers, C. C. (2004). A local
nearest-neighbor convex-hull construction of home ranges
and utilization distributions. _Ecography_, 27(4), 489-505.

Getz, W. M., Fortmann-Roe, S., Cross, P. C., Lyons, A. J.,
Ryan, S. J., & Wilmers, C. C. (2007). LoCoH: nonparameteric
kernel methods for constructing home ranges and utilization
distributions. _PloS one_, 2(2), e207.
}

