% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrMapFields.R
\name{rhrMapFields}
\alias{rhrMapFields}
\title{Map information to columns}
\usage{
rhrMapFields(dat, fields = list(lon = NA, lat = NA, id = NA, date = NA, time =
  NA), projString = NULL, projStringOut = NULL, dateFormat = "ymd",
  timeFormat = "hms", defaultId = "Animal_1")
}
\arguments{
\item{dat}{A data.frame or SptialPointsDataFrame.}

\item{fields}{A list with named elements for lon, lat and id. Optionally also for date and time.}

\item{projString}{A object of class CRS.}

\item{projStringOut}{A object of class CRS, to which the data is transformed to.}

\item{dateFormat}{A character indicating date format, see details.}

\item{timeFormat}{A character indicating time format, see details.}

\item{defaultId}{A character indicating default ID in case id is missing from fields.}
}
\value{
A list of class \code{RhrMappedData}. The list contains the following elements: \code{dat} (SpatialPointsDataFrame, with the data), \code{hasTS} (logical scalar indicating if there is a timestamp) and a list that gives information about the number of missing and duplicated points.
}
\description{
Convenience function to prepare data for further analysis.
}
\details{
For date and time the lubridate conventions are used. E.g. day-month-year is abbreviated as \code{dmy}.
}

