% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrMCP}
\alias{rhrMCP}
\title{rhrMCP}
\usage{
rhrMCP(xy, levels = 95)
}
\arguments{
  \item{xy}{data.frame with two columns. The first column
  contains x coordinates and the second column contains y
  coordinates}

  \item{levels}{vector with the percentage of closest
  points to the centroid that are used to calculated MCP}

  \item{proj4string}{character, with the projection (it is
  parsed with \code{sp::CRS}).}
}
\value{
object of class \code{RhrMCP}
}
\description{
Minimum Convex Polygon (MCP)
}
\details{
Computes the minimum convex polygon of a subset of points.
First the centroid of the home range is found with
\code{rgeos::gCentroid} and then the `100 - levels` points
are used to calculate a minimum convex polygon.
}
\examples{
data(datSH)
## Calculate mcp at one level
mcp1 <- rhrMCP(datSH[, 2:3], levels=95)
## Calculate mcp at several levels
mcp2 <- rhrMCP(datSH[, 2:3], levels=c(50, 90, 95))

## Area at each isopleth level
rhrArea(mcp2)

## SptialPolygonsDataFrame of isopleth
rhrIsopleths(mcp2)
}
\author{
Johannes Signer inspired from \code{adehabitatHR::mcp}
}
\seealso{
\code{adehabitatHR::mcp}, \code{rgeos::gConvexHull}
}

