% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrMCP.R
\name{rhrMCP}
\alias{rhrMCP}
\title{Minimum Convex Polygon (MCP)}
\usage{
rhrMCP(xy, levels = 95)
}
\arguments{
  \item{xy}{Instance of \code{data.frame},
  \code{SpatialPoints}, \code{SpatialPointsDataFrame}
  or \code{RhrMappedData}. If \code{xy} is is a
  \code{data.frame}, then the first two columns are
  assumed to be the \code{x} and \code{y} coordinates of
  the relocations.} \item{levels}{Numeric \code{vector}
  indicating percentages at which the home range estimates
  should be evaluated.}
}
\value{
Object of class \code{RhrMCP}.
}
\description{
Function to estimate the minimum convex polygon (MCP) home
range.
}
\details{
Computes the minimum convex polygon of a subset of
points. First the centroid of the home range is found with
\code{rgeos::gCentroid} and then the `%-levels` points are
used to calculate a minimum convex polygon.
}
\examples{
library(rhr)
data(datSH)

## MCP from data.frame
mcp1 <- rhrMCP(datSH[, 2:3], levels=95)

## Calculate mcp at several levels
mcp2 <- rhrMCP(datSH[, 2:3], levels=c(50, 90, 95))

## Area at each isopleth level
rhrArea(mcp1)
rhrArea(mcp2)

## plot data
plot(mcp1)

## SptialPolygonsDataFrame of isopleth
plot(rhrIsopleths(mcp2))

## Data: not projected
datrhr1 <- rhrMapFields(datSH, fields=list(lon="x_epsg31467", lat="y_epsg31467"))

## Data: projected
datrhr2 <- rhrMapFields(datSH, fields=list(lon="x_epsg31467", lat="y_epsg31467"),
                    projString=sp::CRS("+init=epsg:31467"))

datSP1 <- sp::SpatialPoints(datSH[, 2:3])

mcp4 <- rhrMCP(datrhr1)
mcp5 <- rhrMCP(datrhr2)
mcp6 <- rhrMCP(datSP1)

sapply(list(mcp1, mcp4, mcp5, mcp6), rhrArea)
}
\seealso{
\code{adehabitatHR::mcp}, \code{rgeos::gConvexHull}
}

