% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrN}
\alias{rhrN}
\title{Number of relocations in a track.}
\usage{
rhrN(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{...}{none implemented.}
}
\description{
Returns the number of relocations in a track.
}

