% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrN}
\alias{rhrN}
\title{Number of relocations in a track.}
\usage{
rhrN(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{...}{none implemented.}
}
\description{
Returns the number of relocations in a track.
}
\examples{
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- lubridate::ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackS (only space)
trackS <- rhrTrack(sp)
rhrN(trackS) 

## This the same as
nrow(trackS)

}

