% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrNSD.R
\name{rhrNSD}
\alias{rhrNSD}
\title{Net squared deplacement}
\usage{
rhrNSD(x, ...)
}
\arguments{
\item{x}{RhrTrack* object}
}
\description{
Calcualtes the net squuared deplacement from the first point.
}
\examples{
data(trackS)
traj <- rhrTraj(SpatialPoints(datSH[, 2:3]))
plot(rhrNSD(traj), type= "l")
}

