% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrPoints}
\alias{rhrPoints}
\title{Extract relocations from a track.}
\usage{
rhrPoints(x, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}
}
\description{
Extract the relocation points from a track and return them as a `SpatialPointsPointsDataframe` object.
}
\examples{
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- lubridate::ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackS (only space)
trackS <- rhrTrack(sp)
rhrPoints(trackS)

}

