% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rhrRasterFromExt}
\alias{rhrRasterFromExt}
\title{Create a raster}
\usage{
rhrRasterFromExt(ext, nrow = NULL, ncol = NULL, res = 1, sameRes = TRUE)
}
\arguments{
  \item{ext}{RhrRasterExt, giving the extent of the raster}

  \item{nrow}{numeric; number of rows}

  \item{ncol}{numeric; number of columns}

  \item{res}{numeric; the resolution}
}
\value{
raster
}
\description{
Create a raster layer from extent
}
\details{
This function wraps \code{raster::rast} and adds
functionality to a take control over the resolution.
Resolution is given precedence over \code{nrow} and
\code{ncol}.
}
\examples{
dat <- data.frame(x=rnorm(100), y=rnorm(100))
## Get the bounding box
ext <- rhrExtFromPoints(dat)
r <- rhrRasterFromExt(ff, nrow=10, ncol=100)
}
\author{
Johannes Signer
}

