% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rhrReadData}
\alias{rhrReadData}
\title{read separator delimited text files}
\usage{
rhrReadData(file, sep = ",", skip = 0, hasHeader = TRUE, sepDec = ".",
  stringsAsFactors = FALSE, ...)
}
\arguments{
  \item{file}{character; the file name}

  \item{sep}{character; the separator}

  \item{skip}{numeric; the number of lines to skip}

  \item{hasHeader}{logical; indicates whether or not a
  header is presence}

  \item{sepDec}{character; indicating the decimal
  separator}

  \item{...}{passed to scan}
}
\value{
\code{data.frame}
}
\description{
Wrapper around \code{readLines} to read data with different
field separator and to do some checking, mainly used for
the GUI
}
\details{
some details
}
\author{
Johannes Signer
}

