% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrReadData}
\alias{rhrReadData}
\title{read separator delimited text files}
\usage{
rhrReadData(file, sep = ",", skip = 0, hasHeader = TRUE, sepDec = ".",
  stringsAsFactors = FALSE, ...)
}
\arguments{
  \item{file}{character; the file name}

  \item{sep}{character; the separator}

  \item{skip}{numeric; the number of lines to skip}

  \item{hasHeader}{logical; indicates whether or not a
  header is presence}

  \item{sepDec}{character; indicating the decimal
  separator}

  \item{...}{passed to scan}
}
\value{
object of class \code{rhrDat}
}
\description{
Wrapper around \code{readLines} to read data with different
field separator and to do some checking, mainly used for
the GUI
}
\details{
...
}
\examples{
## Generate data
tmpDir <- tempdir()
dat <- data.frame(x=runif(1000),
                  y=runif(1000))

write.csv(dat, file.path(tmpDir, "test.csv"))

## Read data
dat <- rhrReadData(file.path(tmpDir, "test.csv"), sep=",")
}
\author{
Johannes Signer
}

