% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrRW.R
\name{rhrRW}
\alias{rhrRW}
\title{Simulate a random walk in 2D.}
\usage{
rhrRW(xy0 = c(0, 0), n = 1000, sigma = 1, time)
}
\arguments{
\item{xy0}{Numeric vector of length two. The initial coordinates.}

\item{n}{Numeric scalar. The number of steps.}

\item{sigma}{Numeric scalar or vector of length \code{n}. Controlls the step lengths.}

\item{time}{Vector of \code{POSIXct} objects. Providing the time stamp of relocations. If \code{time} is missing, a sequence of time stamps generated starting at the current time in one minute time steps.}
}
\value{
Object of class RhrTrajST
}
\description{
Individum moves in a planar coordinate system according to a simple random walk.
}
\examples{
set.seed(123)

## All steps with equal sigma
walk <- rhrRW()
plot(walk)

hist(rhrSegments(walk)$distance)

## Three different sigmas
walk <- rhrRW(n = 3000, sigma = rep(c(1, 5, 1), each = 1000))
plot(walk)
hist(rhrSegments(walk[1:1000, ])$distance)
hist(rhrSegments(walk[1001:2000, ])$distance)
hist(rhrSegments(walk[2001:3000, ])$distance)

## Sigma varies over time following a sine curve
sigma <- sin(seq(0, pi, length.out = 1000)) * 2
plot(sigma, type = "l")
walk <- rhrRW(n = 1000, sigma = sigma)
plot(walk)

## Does home-range size change over time as well?
# b <- rhrSplit(walk, rep(1:5, each = 200))
# plot(sapply(b, function(x) rhrArea(rhrMCP(x))$area), 
#     xlab = "time", ylab = "Home-range size")

## Time steps do not have to be equal
times <- lubridate::now() + lubridate::minutes(cumsum(rpois(1000, 10)))
walk <- rhrRW(n = 1000, time = times)
plot(walk)
hist(rhrSegments(walk)$duration)

}

