% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrSegments}
\alias{rhrSegments}
\title{Extract segments from a track.}
\usage{
rhrSegments(x, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}

\item{spatial}{Logical value, should the result be spatial (SpatialLinesDataFrame).}
}
\description{
Extract the segments, linear interpolations between two points, from the track.
}
\examples{
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- lubridate::ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackS (only space)
trackS <- rhrTrack(sp)
class(trackS)

## Create an object of RhrTrackST (only space)
trackST <- rhrTrack(sp, time)
class(trackST)
}

