% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrSiteFidelity.R
\name{rhrSiteFidelity}
\alias{rhrSiteFidelity}
\title{Test for site fidelity of animal movement.}
\usage{
rhrSiteFidelity(dat, n = 100, alpha = 0.05)
}
\arguments{
\item{n}{Numeric scalar. The number of simulated trajectories.}

\item{alpha}{Numeric scalar. The alpha value used for the bootstrapping.}

\item{xy}{Instance of \code{data.frame}, \code{SpatialPoints},
\code{SpatialPointsDataFrame} or \code{RhrMappedData}. If \code{xy} is
is a \code{data.frame}, then the first two columns are assumed to be the
\code{x} and \code{y} coordinates of the relocations.}
}
\value{
object of class \code{RhrFidelity}, which is a list of length 4. \code{msd.dat} and \code{li.dat} is the mean square distance and linearity for the real date. \code{msd.sim} and \code{li.sim} are the mean square distances and linearities for the simulated trajectories.
}
\description{
Test for site fidelity of animal movement.
}
\examples{
set.seed(123)


## Simulated data

# Simulate a random walk.
a <- rhrRW(n = 1000)
plot(a)

# Calcualte site fidelity
sf <- rhrSiteFidelity(a, n = 200)

# For MSD and LI the observed data do not differ significantely from random permutations. 
sf
plot(sf)

# Simulate trajectory as Ornstein-Uhlenbeck process
a <- rhrOU(n = 10000, A = matrix(c(0.1, 0, 0, 0.1), 2))
plot(a)
sf <- rhrSiteFidelity(a, n = 200)

# For MSD and LI the observed data differ significantely from random permutations. 
sf
plot(sf)

## real data
data(datSH)
res <- rhrSiteFidelity(datSH[, 2:3], n = 200)
res
plot(res)
}
\references{
Spencer, S. R., Cameron, G. N., & Swihart, R. K. (1990). Operationally defining home range: temporal dependence exhibited by hispid cotton rats. Ecology, 1817-1822.
}

