% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrSplit}
\alias{rhrSplit}
\title{Split a track}
\usage{
rhrSplit(x, ...)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}

\item{f}{Vector indicating the grouping}

\item{minN}{Integer, the minimum number of relocations required in a group in order to create a new track.}
}
\description{
Returns a list of subtracks.
}
\examples{
library(rhr)
library(lubridate)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- lubridate::ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackST (only space)
trackST <- rhrTrack(sp, time)

## What is the time span?
rhrTrackSpan(trackST)

## split the track by year
tracks <- rhrSplit(trackST, year(rhrTimes(trackST)))
tracks
}

