% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrTimes}
\alias{rhrTimes}
\title{Extract time stamps from a track.}
\usage{
rhrTimes(x, ...)
}
\arguments{
\item{x}{Object of class `RhrTrack*`.}

\item{...}{none implemented.}
}
\description{
Extracts the time stamps from a track and return them as a vector.
}

