% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrack.R
\name{rhrTrack}
\alias{rhrTrack}
\alias{rhrTracks}
\title{Create objects of class `RhrTrack` to represent animal relocation data.}
\usage{
rhrTrack(sp, time, duplicates = "remove", meta)

rhrTracks(sp, ts, id, meta)
}
\arguments{
\item{sp}{SpatialPoints* with the relocation of the animal or \code{RhrMappedData}.}

\item{time}{Vector with objects of class `POSIXct`, giving the time stamp of relocations.}

\item{duplicates}{String, indicating how to handle duplicted time stamps. Currently only  "remove" is supported.}

\item{meta}{Named list, providing meta information about the animal.}

\item{id}{Vector containing the id for each point in \code{sp}.}
}
\description{
Creates object of class `RhrTrack*` from `SpatialPoints*` and, if provided, time stamps.
}
\examples{
library(lubridate)
library(rhr)
data(datSH)

## Create a SpatialPoints objects with the relocation
sp <- sp::SpatialPoints(datSH[, 2:3])

## Parse time
time <- ymd_hms(paste(datSH$day, datSH$time))

## Create an object of RhrTrackS (only space)
trackS <- rhrTrack(sp)
class(trackS)

## Create an object of RhrTrackST (only space)
trackST <- rhrTrack(sp, time)
class(trackST)

## Create a regular trajectory
timeRounded <- round_date(time, "hour")
which6h <- which(hour(timeRounded) == 6)[10:40]
trackSTR <- rhrTrack(sp[which6h, ], timeRounded[which6h])
trackSTR

}

