% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTrackMethods.R
\name{rhrTrackTime}
\alias{rhrTrackBy}
\alias{rhrTrackEnd}
\alias{rhrTrackSpan}
\alias{rhrTrackStart}
\alias{rhrTrackTime}
\alias{rhrTracksEnd}
\alias{rhrTracksSpan}
\alias{rhrTracksStart}
\title{Working with timestamps}
\usage{
rhrTrackStart(x)

rhrTrackEnd(x)

rhrTrackSpan(x)

rhrTracksStart(x)

rhrTracksEnd(x)

rhrTracksSpan(x)

rhrTrackBy(x)
}
\arguments{
\item{x}{Object of class \code{RhrTrack*}}
}
\description{
Functions to deal with time stamps of a track.
}
\details{
`rhrTrackStart` returns the first the time stamp of the beginning of a track. `rhrTrackEnd` returns the last time stamp and `rhrTrackSpan` the time span of a track.
}

