% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrTTSI.R
\name{rhrTTSI}
\alias{rhrTTSI}
\title{Estimate Time To Statistical Independence (TTSI)}
\usage{
rhrTTSI(track, interval, ntimes = 3, ...)
}
\arguments{
\item{track}{Object inheriting from \code{RhrTrackST}.}

\item{interval}{Numeric scalar, initial interval considered in seconds.}

\item{ntimes}{Numeric scalar, the number of times the critical value needs to be exceeded in order to reach independence.}

\item{...}{arguments passed to \code{rhrSchoener}.}
}
\value{
\code{list} with the original data, the intervals considered, if and when time to statistical independence was reached and the call.
}
\description{
This a wrapper around \code{rhrSchoener} to calculate time to statistical indpendence from a series of intervals.
}
\references{
Swihart, R. and Slade N. 1985, Testing for indpendence of observations in animal movement, Ecology, 66(4), 1176 - 1184
}
\seealso{
\code{\link{rhrTrack}} for fruther details and examples on how to create a \code{RhrTrack*} object.
}

