% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrTTSI}
\alias{rhrTTSI}
\title{rhrTTSI
This a wrapper around \code{rhrSchoener} to calculate time to statistical indpendence from a series of intervals.}
\usage{
rhrTTSI(dat, time, interval, ntimes = 3, ...)
}
\arguments{
  \item{dat}{data.frame, with three columns: x and y
  coordinates and a timestamp in seconds.}

  \item{time}{vector with time stamp for each relocation.}

  \item{interval}{numeric value, the initial interval
  considered}

  \item{ntimes}{numeric value, the number of times the
  critical value needs to be exceeded in order to reach
  independence.}

  \item{...}{further arguments passed to
  \code{rhrSchoener}.}
}
\value{
\code{vector} vector of length three. V is Schoeners v, n
the number of points and m the number of point pairs
considered to calculated t2.
}
\description{
rhrTTSI This a wrapper around \code{rhrSchoener} to
calculate time to statistical indpendence from a series of
intervals.
}
\author{
Johannes Signer
}
\references{
Swihart, R. and Slade N. 1985, Testing for indpendence of
observations in animal movement, Ecology, 66(4), 1176 -
1184
}

