% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrUD}
\alias{rhrUD}
\title{Generic Function to calculate UD}
\usage{
rhrUD(x, ...)
}
\arguments{
  \item{x}{an object of class RhrHrEstimator*}

  \item{...}{further arguments, none implemented}
}
\value{
an object of class raster
}
\description{
Generic Function to calculate UD
}

