% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhrUD.R
\name{rhrUD}
\alias{rhrUD}
\title{Utilization distribution}
\usage{
rhrUD(x, ...)
}
\arguments{
\item{x}{An object of class RhrProbEst.}

\item{...}{none implemented.}
}
\value{
An object of class RasterLayer.
}
\description{
Function to obtain the utilization distribution (UD) of a home range estimate.
}
\details{
For probabilistic estimators home range estimator this functions returns the UD.
}

