% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rhrUniNorm}
\alias{rhrUniNorm}
\title{rhrUniNorm}
\usage{
rhrUniNorm(xy, trast = rhrRasterFromExt(rhrExtFromPoints(xy, extendRange =
  0.2), nrow = 100, res = NULL))
}
\arguments{
  \item{xy}{valid input data}

  \item{trast}{template raster}
}
\description{
Univariate Normal Home-Range estimation
}
\details{
Computes home-range using univariate normal distribution
}

